pyproject.toml
[project]
name = "mcp-devsecops-server"
version = "0.2.0"
description = "MCP server para orquestar pipelines DevSecOps"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.27.1",
    "python-dotenv>=1.2.2",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"